docs: enhanced sidebar context
