+++ www/docs/src/theme/DocPage/index.tsx
@@ -17,6 +17,7 @@ import NotFound from "@theme/NotFound"
 import SearchMetadata from "@theme/SearchMetadata"
 import type { Props } from "@theme/DocPage"
 import { SidebarContext } from "@site/src/context/sidebar"
+import SidebarProvider from "@site/src/providers/Sidebar"
 
 function DocPageMetadata(props: Props): JSX.Element {
   const { versionMetadata } = props
@@ -46,53 +47,6 @@ export default function DocPage(props: Props): JSX.Element {
   }
   const { docElement, sidebarName, sidebarItems } = currentDocRouteMetadata
 
-  const [hiddenSidebar, setHiddenSidebar] = useState(false)
-  const [hiddenSidebarContainer, setHiddenSidebarContainer] = useState(false)
-  const [floatingSidebar, setFloatingSidebar] = useState(false)
-  const toggleSidebar = useCallback(() => {
-    if (hiddenSidebar) {
-      setHiddenSidebar(false)
-    }
-    // onTransitionEnd won't fire when sidebar animation is disabled
-    // fixes https://github.com/facebook/docusaurus/issues/8918
-    if (!hiddenSidebar && prefersReducedMotion()) {
-      setHiddenSidebar(true)
-    }
-    setHiddenSidebarContainer((value) => !value)
-  }, [setHiddenSidebarContainer, hiddenSidebar])
-
-  useEffect(() => {
-    function isEditingContent(event: KeyboardEvent) {
-      const element = event.target as HTMLElement
-      const tagName = element.tagName
-      return (
-        element.isContentEditable ||
-        tagName === "INPUT" ||
-        tagName === "SELECT" ||
-        tagName === "TEXTAREA"
-      )
-    }
-
-    function sidebarShortcut(e: KeyboardEvent) {
-      if (
-        (e.metaKey || e.ctrlKey) &&
-        e.key.toLowerCase() === "i" &&
-        !isEditingContent(e)
-      ) {
-        e.preventDefault()
-        toggleSidebar()
-      }
-    }
-
-    window.addEventListener("keydown", sidebarShortcut)
-
-    return () => {
-      window.removeEventListener("keydown", sidebarShortcut)
-    }
-  })
-
-  console.log(sidebarName)
-
   return (
     <>
       <DocPageMetadata {...props} />
@@ -107,20 +61,9 @@ export default function DocPage(props: Props): JSX.Element {
       >
         <DocsVersionProvider version={versionMetadata}>
           <DocsSidebarProvider name={sidebarName} items={sidebarItems}>
-            <SidebarContext.Provider
-              value={{
-                hasSidebar: sidebarName !== undefined,
-                hiddenSidebar,
-                setHiddenSidebar,
-                hiddenSidebarContainer,
-                setHiddenSidebarContainer,
-                floatingSidebar,
-                setFloatingSidebar,
-                onCollapse: toggleSidebar,
-              }}
-            >
+            <SidebarProvider sidebarName={sidebarName}>
               <DocPageLayout>{docElement}</DocPageLayout>
-            </SidebarContext.Provider>
+            </SidebarProvider>
           </DocsSidebarProvider>
         </DocsVersionProvider>
       </HtmlClassNameProvider>
